runner: treat command timeout as a failed check
there are several levels of errors when runner executes checks. The error
is considered to be an exception, where runner is unable to run the checks for
some reason. For example, when the check is not present in the path.
However if the command execution takes more time then defined in `timeout`
check property, it will be killed and also considered as an error. But in fact,
it should be a failed check. This PR fixes it.