Fix broken logic in tests
PRevious commit f091937393eea045d90c92be59abacdfb436e252 introduced
checking for bundle clomplete. Unfortunatelly it was not a real fix
since all requests were mocked. This PR changes this logic to use
filesystem instead of usign API. In the future we should add integration
test in for this project where we could test whole flow and interact
with API instead of filestystem.