Fix compilation on Darwin by adding noop stubs of interfaces
This adds the necessary noop functions to allow compilation on Darwin,
which makes it possible to run tests/linters/etc.

When possible I tried to copy the behaviour of the
Windows-implementations and where not I opted to return an error to
signal that it's an unsupported feature.

I'm still unsure about the comments.