Introduce featcher (#91)
Fetcher will download endpoints that it receive from `input` channel. After download in the response body will be saved into a local zip file for every fetcher instance. Updates will be published on `statusUpdate` channel. When input channel is closed or context is cancelled fetcher will post the path to the zip file on response channel.