Enable optional files in diagnsotics-bundle
Introduce `Optional` option in endpoint configuration for files.
If a file is marked as optional then errors are swallowed and do not affect bundle creation process.
Error is included as file content. This will allow us to getcher files
that are not always available on cluster such as rotated log files or
configuration files for optional components.

Fixes: https://jira.mesosphere.com/browse/DCOS_OSS-4949