Allow optional commands
Currently if command does not exists we create empty file.
This change includes stderr in command output file and
if command does not exists or error (returns non 0 result) then
makr it as failed and note it in summarryErrorReport.txt.
This change introduce also `Optional` field for command so
event if they fail, the bundle could be finished whitout errors
just like files.

Fixes: DCOS_OSS-5026
Refs: https://github.com/dcos/dcos-diagnostics/pull/94