Added timeout to systemd log reading; fixed `verbose` flag
This closes https://jira.mesosphere.com/browse/DCOS_OSS-5097

I added a TimeoutReadCloser to our units package that reads data from
an underlying ReadCloser only when the deadline of the given Context
has not been reached. Otherwise it stops reading and returns the error
retrieved from the Context (usually a DeadlineExceeded error).

When reading a systemd unit's log is interrupted we now log this fact
in the diagnostics output so that the consumer of the bundle knows
there's probably more in the journal than he sees.